Regroup and REOrganize of service folder
