+++ server/services/chat_models/openai_response.py
